Added delete item confirmation

+++ frontend/src/app/item/item.component.ts
@@ -33,15 +33,18 @@ export class ItemComponent implements OnInit {
   }
 
   deleteItem(id: number) {
-    this.itemService.deleteItem(id).subscribe(  // The component subscribes to this observable to delete item; Calling deleteItem from item.service.ts
-      {
-        next: () => {
-          this.items = this.items.filter(item => item.id !== id);   // Removing deleted item from the local items array
-        },
-        error: (error) => {
-          console.error('Error deletung item', error);
+    const confirmed = confirm(`Are you sure you want to delete the item?`);
+    if (confirmed) {
+      this.itemService.deleteItem(id).subscribe(  // The component subscribes to this observable to delete item; Calling deleteItem from item.service.ts
+        {
+          next: () => {
+            this.items = this.items.filter(item => item.id !== id);   // Removing deleted item from the local items array
+          },
+          error: (error) => {
+            console.error('Error deletung item', error);
+          }
         }
-      }
-    );
+      );
+    }
   }
 }
